Merge disk.proto into client.proto.
This change is needed to use java code generator. In GO proto gen mode, we use bash to go inside each folder
and generate go code. Sort of like:
```sh
cd kube/v1beta2; protoc -out= *.proto
```

In Java mode, we are using official gradle plugin. Gradle plugin does it's own recursive traversing. So, it expects paths from the root
of the repo. In that case, we also have to specify them by their pkg path.

So, the easy fix is to not use multiple files in a folder, if they need to refer each other.